requirements.txt
requests
pillow
ebooklib